requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-multipart==0.0.9
